+++ functions.php
@@ -34,7 +34,8 @@ require_once 'inc/class/class-svg-icons.php';
  */
 require_once 'inc/helpers/wp-cleanup.php'; // Remove unneeded WordPress default functionality.
 require_once 'inc/helpers/svg-icons.php';
-require_once 'inc/helpers/template-tags.php';
+require_once 'inc/helpers/featured-images.php';
+require_once 'inc/helpers/helper-functions.php';
 
 /**
  * Block renders